refactor: 컴포넌트 이름 변경 - Edit와 관련된 컴포넌트임을 명확하게 보여주는 이름으로 변경
