+++ src/components/Editor/SubDocumentEditor.js
